Rename folder resource
