+++ src/flat/folders/simple/node.folder.image.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
 import * as osc from "outscale-api";
-import { ExplorerNode, ExplorerFolderNode, Profile } from './node';
-import { FolderNode } from './node.folder';
-import { ResourceNode } from './node.resources';
-import { deleteOMI, getOMIs } from '../cloud/omis';
-import { getAccounts } from '../cloud/account';
+import { ExplorerNode, ExplorerFolderNode, Profile } from '../../node';
+import { FolderNode } from '../node.folder';
+import { ResourceNode } from '../../resources/node.resources';
+import { deleteOMI, getOMIs } from '../../../cloud/images';
+import { getAccounts } from '../../../cloud/account';
 
 export class OMIsFolderNode extends FolderNode implements ExplorerFolderNode {
     constructor(readonly profile: Profile) {
