+++ src/flat/folders/simple/node.folder.routetable.ts
@@ -1,8 +1,8 @@
 import * as vscode from 'vscode';
-import { ExplorerNode, ExplorerFolderNode, Profile } from './node';
-import { FolderNode } from './node.folder';
-import { ResourceNode } from './node.resources';
-import { deleteRouteTable, getRouteTables } from '../cloud/routetables';
+import { ExplorerNode, ExplorerFolderNode, Profile } from '../../node';
+import { FolderNode } from '../node.folder';
+import { ResourceNode } from '../../resources/node.resources';
+import { deleteRouteTable, getRouteTables } from '../../../cloud/routetables';
 
 export class RouteTablesFolderNode extends FolderNode implements ExplorerFolderNode {
     constructor(readonly profile: Profile) {
